Implement #FP20130311 Setting for deletions. If for example I delete synced file from the cloud so that the deletion will not happen in the phone. And possibility to select scheduled one-way sync ( to local or to remote) or two-way sync. -Relate to #FP20130297 -Currently, deleted cloud-connected file will always be deleted locally. Delete connected local file will also delete on cloud. (2-ways) -Options: SyncBoth SyncForward == forcePut SyncBackward == forceGet -Please review cases in sheet SyncDirections cases." (WIP)

+++ qml/FilesPlus/meego/main.qml
@@ -366,7 +366,11 @@ PageStackWindow {
                 if (cloudDriveModel.isConnected(sourcePath)) {
                     var json = Utility.createJsonObj(cloudDriveModel.getItemListJson(sourcePath));
                     for (var i=0; i<json.length; i++) {
-                        cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        if (json[i].sync_direction === CloudDriveModel.SyncBackward) {
+                            cloudDriveModel.disconnect(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path);
+                        } else {
+                            cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        }
                     }
 
                     // Reset cloudDriveModel hash on parent. CloudDriveModel will update with actual hash once it got reply.
@@ -490,7 +494,11 @@ PageStackWindow {
                 if (cloudDriveModel.isConnected(sourcePath)) {
                     var json = Utility.createJsonObj(cloudDriveModel.getItemListJson(sourcePath));
                     for (var i=0; i<json.length; i++) {
-                        cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        if (json[i].sync_direction === CloudDriveModel.SyncBackward) {
+                            cloudDriveModel.disconnect(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path);
+                        } else {
+                            cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        }
                     }
 
                     // Reset cloudDriveModel hash on parent. CloudDriveModel will update with actual hash once it got reply.
