+++ qml/FilesPlus/symbian/main.qml
@@ -370,7 +370,11 @@ PageStackWindow {
                 if (cloudDriveModel.isConnected(sourcePath)) {
                     var json = Utility.createJsonObj(cloudDriveModel.getItemListJson(sourcePath));
                     for (var i=0; i<json.length; i++) {
-                        cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        if (json[i].sync_direction === CloudDriveModel.SyncBackward) {
+                            cloudDriveModel.disconnect(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path);
+                        } else {
+                            cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        }
                     }
 
                     // Reset cloudDriveModel hash on parent. CloudDriveModel will update with actual hash once it got reply.
@@ -494,7 +498,11 @@ PageStackWindow {
                 if (cloudDriveModel.isConnected(sourcePath)) {
                     var json = Utility.createJsonObj(cloudDriveModel.getItemListJson(sourcePath));
                     for (var i=0; i<json.length; i++) {
-                        cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        if (json[i].sync_direction === CloudDriveModel.SyncBackward) {
+                            cloudDriveModel.disconnect(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path);
+                        } else {
+                            cloudDriveModel.deleteFile(json[i].type, json[i].uid, json[i].local_path, json[i].remote_path, true); // NOTE suppressDeleteLocal=true
+                        }
                     }
 
                     // Reset cloudDriveModel hash on parent. CloudDriveModel will update with actual hash once it got reply.
